+++ naming/src/main/java/com/alibaba/nacos/naming/consistency/ephemeral/distro/newimpl/component/DistroComponentHolder.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.alibaba.nacos.naming.consistency.ephemeral.distro.newimpl;
+package com.alibaba.nacos.naming.consistency.ephemeral.distro.newimpl.component;
 
 import com.alibaba.nacos.naming.consistency.ephemeral.distro.newimpl.task.delay.DistroDelayTaskExecuteEngine;
 import com.alibaba.nacos.naming.consistency.ephemeral.distro.newimpl.task.execute.DistroExecuteWorkersManager;
