+++ naming/src/main/java/com/alibaba/nacos/naming/consistency/ephemeral/distro/newimpl/component/DistroTransportAgent.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.alibaba.nacos.naming.consistency.ephemeral.distro.newimpl;
+package com.alibaba.nacos.naming.consistency.ephemeral.distro.newimpl.component;
 
 import com.alibaba.nacos.naming.consistency.ephemeral.distro.newimpl.entity.DistroData;
 import com.alibaba.nacos.naming.consistency.ephemeral.distro.newimpl.entity.DistroKey;
