+++ naming/src/main/java/com/alibaba/nacos/naming/consistency/ephemeral/distro/newimpl/task/delay/DistroDelayTaskProcessor.java
@@ -19,7 +19,7 @@ package com.alibaba.nacos.naming.consistency.ephemeral.distro.newimpl.task.delay
 import com.alibaba.nacos.common.task.AbstractDelayTask;
 import com.alibaba.nacos.common.task.NacosTaskProcessor;
 import com.alibaba.nacos.naming.consistency.ApplyAction;
-import com.alibaba.nacos.naming.consistency.ephemeral.distro.newimpl.DistroComponentHolder;
+import com.alibaba.nacos.naming.consistency.ephemeral.distro.newimpl.component.DistroComponentHolder;
 import com.alibaba.nacos.naming.consistency.ephemeral.distro.newimpl.entity.DistroKey;
 import com.alibaba.nacos.naming.consistency.ephemeral.distro.newimpl.task.execute.DistroSyncChangeTask;
