+++ naming/src/main/java/com/alibaba/nacos/naming/consistency/ephemeral/distro/newimpl/task/verify/DistroVerifyTask.java
@@ -19,7 +19,7 @@ package com.alibaba.nacos.naming.consistency.ephemeral.distro.newimpl.task.verif
 import com.alibaba.nacos.core.cluster.Member;
 import com.alibaba.nacos.core.cluster.ServerMemberManager;
 import com.alibaba.nacos.naming.consistency.ApplyAction;
-import com.alibaba.nacos.naming.consistency.ephemeral.distro.newimpl.DistroComponentHolder;
+import com.alibaba.nacos.naming.consistency.ephemeral.distro.newimpl.component.DistroComponentHolder;
 import com.alibaba.nacos.naming.consistency.ephemeral.distro.newimpl.entity.DistroData;
 import com.alibaba.nacos.naming.consistency.ephemeral.distro.newimpl.entity.DistroKey;
 import com.alibaba.nacos.naming.misc.Loggers;
